// src/app/countries/interfaces/cache-storage.interface.ts
import { Country } from './countries.interface';
import { Region } from './region.type';

export interface CacheStoraged {
  byCapital: TermCountries;
  byCountry: TermCountries;
  byRegion: RegionCountries;
}

export interface TermCountries {
  term: string;
  countries: Country[];
}

export interface RegionCountries {
  region?: Region;
  countries: Country[];
}